refactor(ai): split the package by domain and extract prompts
Break the single ai.ts module into one file per concern — types, errors, the
json/severity/diff/gate/assessment/provider helpers, and the reviewer
orchestrator — re-exporting the public surface from mod.ts. Move the prompt
text into src/prompts/ (subjects and templates) so the wording can be edited
without touching the logic. Reuse core's globToRegExp instead of a hand-rolled
glob matcher.

Record the rule in CLAUDE.md: one domain per file, never the whole
implementation in one module, and prefer reusing core primitives.
